Add files via upload
I have managed to include the eyes which move with the mouse pointer. I want to add an extra pupil to make the eye more realistic and contain the eye within the circle as it is coming out right now. How can mix this with the type sampler which changes with each key pressed so that when any letter with the circle is typed (ex R, P, O) the eye movement comes along and then disappears as the next letter is typed